Interval class and Interval object separated.
